Moves posts under the admin namespace
