+++ app/controllers/admin/users_controller.rb
@@ -3,11 +3,9 @@
 class Admin::UsersController < ApplicationController
   include Pagy::Backend
 
-  before_action :authenticate_user!
+  before_action :admin_only!
 
   def index
-    render plain: "404 Not Found", status: :not_found unless admin?
-
     @pagy, @users = pagy(User.order(created_at: :desc))
   end
 end
