+++ app/views/admin/_toolbar.html.erb
@@ -1,5 +1,5 @@
 <nav class="admin-toolbar" aria-label="Admin toolbar">
   <%= link_to "Admin", admin_root_url %> |
-  <%= link_to "New Post", new_post_url %> |
+  <%= link_to "New Post", new_admin_post_url %> |
   <%= link_to "Mailing List", mailing_lists_url %> |
 </nav>
