+++ app/views/admin/admin/index.html.erb
@@ -1,6 +1,6 @@
 <h1>Admin Section</h1>
 <ul>
-  <li><%= link_to "Add a new post", new_post_url %></li>
+  <li><%= link_to "Add a new post", new_admin_post_url %></li>
   <li><%= link_to "View all posts", posts_url %></li>
   <li><%= link_to "View all users", admin_users_url %></li>
 </ul>
