+++ app/views/admin/posts/_form.html.erb
