+++ app/views/posts/index.html.erb
@@ -1,7 +1,7 @@
 <%- if @posts %>
   <hr>
   <%- @posts.each do |post| %>
-    <h2><%= link_to post.title, edit_post_url(post.id) %></h2>
+    <h2><%= link_to post.title, edit_admin_post_url(post.id) %></h2>
     <p><%= post.content %></p>
     <hr>
   <%- end %>
